analyzer: Add a `PackageManagerDependencyHandler`
The `PackageManagerDependencyHandler` can be used to create and resolve
dependencies between different package managers.

Package managers can add a dependency on another definition file and
scope name to their dependency tree using the
`createPackageManagerDependency()` function. This dependency is resolved
when all package managers have finished, by searching for projects from
the named definition file and adding a dependency to each of these
projects, with the transitive dependencies defined by the scope name.

This is implemented as an operation on the dependency graph, to provide
better performance than if it was implemented on the dependency tree.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>